third_party/WebKit/LayoutTests/platform/chromium-win-xp/fast/spatial-navigation/snav-div-overflow-scrol-hidden-expected.txt










PASS gFocusedDocument.activeElement.getAttribute("id") is "f2"
PASS gFocusedDocument.activeElement.getAttribute("id") is "f3"
FAIL gFocusedDocument.activeElement.getAttribute("id") should be f3. Was f4.
FAIL gFocusedDocument.activeElement.getAttribute("id") should be f4. Was f5.
FAIL gFocusedDocument.activeElement.getAttribute("id") should be f4. Was f6.
FAIL gFocusedDocument.activeElement.getAttribute("id") should be f5. Was f9.
FAIL gFocusedDocument.activeElement.getAttribute("id") should be f6. Was f9.
PASS gFocusedDocument.activeElement.getAttribute("id") is "f9"
This test is testing that div with overflow:auto would scroll, by div with overflow:hidden would not.
